Using uint32_t instead of size_t
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@15730 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ src/c/socket/xmlBlasterSocket.c
@@ -536,7 +536,7 @@ Dll_Export MsgUnitArr *parseMsgUnitArr(size_t dataLen, char *data)
 {
    MsgUnitArr *msgUnitArr = (MsgUnitArr *)calloc(1, sizeof(MsgUnitArr));
    size_t currpos = 0;
-   size_t currIndex = 0;
+   uint32_t currIndex = 0;
    enum { SIZE = 56 };
    msgUnitArr->isOneway = false;
    if (dataLen <= 0) {
